#8 refactor(DepartmentNoticeRepository): Long -> Integer 타입으로 변경 및 자바독 내용 변경

+++ src/main/java/me/gijung/DMforU/repository/DepartmentNoticeRepository.java
@@ -9,7 +9,7 @@ public interface DepartmentNoticeRepository extends JpaRepository<DepartmentNoti
 
     /** 학과의 가장 최신 게시글 번호를 확인하는 메서드
      * @Param: 학과 이름
-     * @Return: 해당 학과의 최신 게시글 번호*/
+     * @Return: 해당 학과의 최신 게시글 번호, 만약 게시글이 존재하지 않다면 Null을 반환한다.*/
     @Query("SELECT MAX(e.number) FROM DepartmentNotice e WHERE e.department = :departmentName")
-    Long findMaxNumberByDepartmentName(@Param("departmentName") String departmentName);
+    Integer findMaxNumberByDepartmentName(@Param("departmentName") String departmentName);
 }
